# Conflicts:
#	order-microservice/src/main/java/nl/tudelft/sem/template/user/api/UserMicroServiceAPI.java
#	order-microservice/src/main/java/nl/tudelft/sem/template/user/services/UserMicroServiceService.java

+++ order-microservice/src/main/java/nl/tudelft/sem/template/order/controllers/OrderController.java
@@ -8,6 +8,8 @@ import nl.tudelft.sem.template.model.Address;
 import nl.tudelft.sem.template.model.Order;
 import nl.tudelft.sem.template.order.domain.helpers.FilteringByStatus;
 import nl.tudelft.sem.template.order.domain.helpers.FilteringParam;
+import nl.tudelft.sem.template.order.domain.helpers.OrderValidation;
+import nl.tudelft.sem.template.order.domain.user.CustomerNotFoundException;
 import nl.tudelft.sem.template.order.domain.user.NoOrdersException;
 import nl.tudelft.sem.template.order.domain.user.NullFieldException;
 import nl.tudelft.sem.template.order.domain.user.OrderNotFoundException;
@@ -24,7 +26,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderController implements OrderApi {
 
     private final transient OrderService orderService;
-    public final transient UserMicroServiceService userMicroServiceService;
+    private final transient UserMicroServiceService userMicroServiceService;
     private final transient DishController dishController;
 
     /**
@@ -96,6 +98,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID The provided ID of the order to return
      * @return 200 OK - The provided Order is returned
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         400 BAD REQUEST - Returning the provided Order is unsuccessful
      *         404 NOT FOUND - No Order exists with the provided ID
      */
@@ -115,6 +118,26 @@ public class OrderController implements OrderApi {
     }
 
     /**
+     * Endpoint for returning an order to a vendor by specifying its id.
+     * The order is returned only after it is paid and if all of its dishes are available
+     *
+     * @param orderId the id of the order to be retrieved
+     * @return 200 OK if the order has been successfully retrieved, including the order
+     *         404 NOT FOUND if the order could not be found
+     *         400 BAD REQUEST if the retrieval was not successful or the order was invalid
+     */
+    @Override
+    public ResponseEntity<Order> orderOrderIDVendorGet(UUID orderId) {
+        OrderValidation orderValidation = new OrderValidation(this, dishController);
+        ResponseEntity<Boolean> response = orderValidation.isOrderValid(orderId);
+        if (Boolean.TRUE.equals(response.getBody())) {
+            ResponseEntity<Order> order = getOrderById(orderId);
+            return order;
+        }
+        return ResponseEntity.notFound().build();
+    }
+
+    /**
      * Endpoint for editing an Order.
      *
      * @param orderID The provided ID of the order to be edited
@@ -131,10 +154,17 @@ public class OrderController implements OrderApi {
         }
 
         try {
-            Order edited = orderService.editOrderByID(orderID, order);
-            return ResponseEntity.ok(edited);
-        } catch (NullFieldException e) {
-            return ResponseEntity.unprocessableEntity().build();
+            String jsonUser = userMicroServiceService.getUserInformation(userID);
+            if (jsonUser == null || jsonUser.isEmpty()) {
+                throw new UserIDNotFoundException(userID);
+            }
+            String userType = JsonParserService.parseUserType(jsonUser);
+            if (userType.equals("Admin")) {
+                Order edited = orderService.editOrderByID(orderID, order);
+                return ResponseEntity.ok(edited);
+            } else {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
@@ -153,8 +183,17 @@ public class OrderController implements OrderApi {
     @Override
     public ResponseEntity<Void> deleteOrderByID(UUID orderID, UUID userID) {
         try {
-            orderService.deleteOrderByID(orderID);
-            return ResponseEntity.ok().build();
+            String jsonUser = userMicroServiceService.getUserInformation(userID);
+            if (jsonUser == null || jsonUser.isEmpty()) {
+                throw new UserIDNotFoundException(userID);
+            }
+            String userType = JsonParserService.parseUserType(jsonUser);
+            if (userType.equals("Admin")) {
+                orderService.deleteOrderByID(orderID);
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
@@ -210,6 +249,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be retrieved the list of dishes from.
      * @return 200 OK if the list of dishes has been retrieved, including the list of UUID dishes
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -233,6 +273,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be retrieved the special requirements from.
      * @return 200 OK if the special requirements of the order has been retrieved, including the special requirements
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -255,6 +296,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be delivered to the retrieved address.
      * @return 200 OK if the address of the order has been retrieved, including the address
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -277,6 +319,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order of which the date of creation should be retrieved.
      * @return 200 OK if the date of the order has been retrieved, including the date
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -286,6 +329,32 @@ public class OrderController implements OrderApi {
             return ResponseEntity.ok(order.getDate());
         } catch (NullFieldException e) {
             return ResponseEntity.unprocessableEntity().build();
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**
+     * OrderID getCustomerName controller method.
+     * It returns the name of the customer that created a specific order
+     *
+     * @param orderID the id of the order from which the customer name should be retrieved
+     * @return 200 OK if the name of the customer has been successfully retrieved, including the name
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
+     *         404 NOT FOUND if the order or the user could not be found
+     *         400 BAD REQUEST if the retrieval was not successful
+     */
+    public ResponseEntity<String> getCustomerName(UUID orderID) {
+        try {
+            Order order = orderService.getOrderById(orderID);
+            UUID userID = order.getCustomerID();
+            return ResponseEntity.ok(userMicroServiceService.getUserName(userID));
+        } catch (NullFieldException e) {
+            return ResponseEntity.unprocessableEntity().build();
+        } catch (UserIDNotFoundException userIDNotFoundException) {
+            return ResponseEntity.notFound().build();
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
@@ -309,6 +378,10 @@ public class OrderController implements OrderApi {
             return ResponseEntity.ok(allOrdersByCustomerID);
         } catch (NoOrdersException noOrdersException) {
             return ResponseEntity.notFound().build();
+        } catch (CustomerNotFoundException customerNotFoundException) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
         }
     }
